feat: add auth route
