+++ src/routes/api/login.ts
@@ -16,8 +16,3 @@ export const post: RequestHandler = async ({ request }) => {
 		headers: { 'set-cookie': a }
 	};
 };
-
-// const cookies = cookie.parse(request.headers.cookie || '');
-// let signedCookies = cookieParser.signedCookies(cookies, [COOKIE_SECRET]);
-// signedCookies = cookieParser.JSONCookies(signedCookies);
-// const jwtToken = signedCookies['jwt'];
