+++ src/routes/auth.svelte
@@ -1,0 +1,28 @@
+<script lang="ts">
+	import { goto } from '$app/navigation';
+
+	let loginInput = '';
+
+	const login = async () => {
+		const body = { token: loginInput };
+		const res = await fetch('/api/login', {
+			method: 'POST',
+			body: JSON.stringify(body),
+			headers: { 'content-type': 'application/json' }
+		});
+		if (res.ok) {
+			goto('/');
+		}
+	};
+</script>
+
+<div class="min-h-screen bg-cyan-400 flex flex-col items-center justify-center">
+	<input
+		placeholder="Login"
+		id="login-input"
+		class="mb-4 p-2 rounded-lg outline-none focus:border-rose-500 border-2"
+		type="text"
+		bind:value={loginInput}
+	/>
+	<button class="bg-cyan-500 p-2 rounded-lg" on:click={() => login()}>Login</button>
+</div>
